수정한 테이블에 맞게 vo 수정
